// prism-desktop/frontend/types/nle.ts
export interface BezierCP {
  x1: number;
  y1: number;
  x2: number;
  y2: number;
}

export interface Keyframe {
  t: number; // time in seconds relative to clip start
  v: number;
  interpolation: 'linear' | 'ease-in' | 'ease-out' | 'ease-in-out' | 'bezier';
  bezierCP?: BezierCP;
}